Update UseraddedtoPrivilgedGroups.yaml to include values for User
Added case for InitiatedbyDisplayname and extended values for if a user implemented the change (Looked pointless to use case for everything as the only other thing that might be the same is the principal name) The IPAddress as well as the display name isn't always populated, but I believe it is important for this rule. 

Please do not hesitate to correct my approach here if it is not efficient.