Renamed to azuredeploy.json
Was receiving this error: There was an error downloading the template from URI 'https://raw.githubusercontent.com/Azure/Azure-Sentinel/master/Playbooks/Get-SOCActions/azuredeploy.json'. Ensure that the template is publicly accessible and that the publisher has enabled CORS policy on the endpoint. To deploy this template, download the template manually and paste the contents in the 'Build your own template in the editor' option below.

Renamed to address the error and allow for deployment.